requirements.txt
Flask
python-dotenv
google-cloud-aiplatform
google-cloud-texttospeech
transformers
torch
gunicorn
